refactored GetPath method to its own class
